feat:  Reconstruct the creation logic of the main window and add a window management module
